+++ upsetplot/plotting.py
@@ -44,7 +44,9 @@ def _process_data(df, *, sort_by, sort_categories_by, subset_size,
     # XXX: ugly!
     def _pack_binary(X):
         X = pd.DataFrame(X)
-        out = 0
+        # use objects if arbitrary precision integers are needed
+        dtype = np.object_ if X.shape[1] > 62 else np.uint64
+        out = pd.Series(0, index=X.index, dtype=dtype)
         for i, (_, col) in enumerate(X.items()):
             out *= 2
             out += col