+++ upsetplot/reformat.py
@@ -288,19 +288,18 @@ def query(data, present=None, absent=None,
     >>>
     >>> # Getting each subset's data
     >>> result = query(data)
-    >>> result.subsets
-    {frozenset({'cat1', 'cat2'}):
-                    index     value
-        cat1  cat2 cat0
-        False True False      3  1.333795,
-    frozenset({'cat1'}):
-                            index     value
-        cat1  cat2  cat0
-        False False False      5  0.918174
-                    False      8  1.948521
-                    False      9  1.086599
-                    False     13  1.105696
-                    False     19  1.339895}
+    >>> result.subsets[frozenset({"cat1", "cat2"})]
+                index     value
+    cat1  cat2 cat0
+    False True False      3  1.333795
+    >>> result.subsets[frozenset({"cat1"})]
+                        index     value
+    cat1  cat2  cat0
+    False False False      5  0.918174
+                False      8  1.948521
+                False      9  1.086599
+                False     13  1.105696
+                False     19  1.339895
     """
 
     data, agg = _aggregate_data(data, subset_size, sum_over)
@@ -332,7 +331,8 @@ def query(data, present=None, absent=None,
 
     # sort:
     if sort_categories_by in ('cardinality', '-cardinality'):
-        totals.sort_values(ascending=sort_categories_by[:1] == '-', inplace=True)
+        totals.sort_values(ascending=sort_categories_by[:1] == '-',
+                           inplace=True)
     elif sort_categories_by is not None:
         raise ValueError('Unknown sort_categories_by: %r' % sort_categories_by)
     data = data.reorder_levels(totals.index.values)