+++ upsetplot/tests/test_upsetplot.py
@@ -289,6 +289,7 @@ def test_include_empty_subsets():
                         common_intersections)
     include_empty_upset.plot()  # smoke test
 
+
 @pytest.mark.parametrize('kw', [{'sort_by': 'blah'},
                                 {'sort_by': True},
                                 {'sort_categories_by': 'blah'},
@@ -940,3 +941,15 @@ def test_style_subsets_artists(orientation):
 
     # TODO: check lines between dots
     # matrix_line_collection = upset_axes["matrix"].collections[1]
+
+
+def test_many_categories():
+    # Tests regressions against GH#193
+    n_cats = 250
+    index1 = [True, False] + [False] * (n_cats - 2)
+    index2 = [False, True] + [False] * (n_cats - 2)
+    columns = [chr(i + 33) for i in range(n_cats)]
+    data = pd.DataFrame([index1, index2], columns=columns)
+    data["value"] = 1
+    data = data.set_index(columns)["value"]
+    UpSet(data)