+++ CHANGELOG.rst
@@ -3,6 +3,8 @@ What's new in version 0.8
 
 - Allowed ``show_percentages`` to be provided with a custom formatting string,
   for example to show more decimal places. (:issue:`194`)
+- Added `include_empty_subsets` to `UpSet` and `query` to allow the display of
+  all possible subsets. (:issue:`185`)
 - Added `subsets` attribute to QueryResult. (:issue:`198`)
 
 What's new in version 0.7